pr-review-service/requirements.txt
agentops>=0.3.0
openai>=1.0.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0